commomFunction.js에서 시간 관리 함수 따로 파일로 분리
