+++ server/DELETE.js
@@ -19,18 +19,13 @@ function deleteStatusOnServer(statusIndex) {
     // 서버에서 status 데이터 삭제
     fetch(FETCH_STATUS_URL + "/" + statusIndex, {
         method: DELETE_METHOD
-    }).then((res) => { 
-        if(res.status != HTTP_SUCCESS) {
-        // 에러 상황 출력을 위해 임시로 해둠
-        // 추후에 에러 핸들링해주기
-        throw new Error(`error : ${statusIndex}번의 card를 삭제하는데 실패`)
-    } })
+    })
 }
 
 /** status를 삭제합니다. */
 function deleteStatus(statusName) {
     let filteredStatus = statusListOnLocal
-                        .filter((statusJSON) => statusJSON[STATUS_NAME] == statusName)[0];
+    .filter((statusJSON) => statusJSON[STATUS_NAME] == statusName)[0];
 
     let statusIndex = filteredStatus[STATUS_INDEX];
 
@@ -50,11 +45,6 @@ function deleteCardDataOnServer(cardID) {
     fetch(FETCH_CARD_URL + "/" + cardID, {
         method: DELETE_METHOD,
     })
-    .then((res) => { if(res.status != HTTP_SUCCESS) {
-        // 에러 상황 출력을 위해 임시로 해둠
-        // 추후에 에러 핸들링해주기
-        throw new Error(`error : ${cardID}번의 card를 삭제하는데 실패`)
-    } })
 }
 
 /** 카드 JSON 데이터를 삭제합니다. */
